modif au niveau des parametre des class int en double et ajout de la class cercle

+++ src/main/java/org/example/Carre.java
@@ -4,7 +4,7 @@ public class Carre implements Forme{
     String Nom;
     Point p1;
     double cote;
-    public Carre(String nom ,double cote , int x , int y){
+    public Carre(String nom ,double cote , double x , double y){
         this.Nom=nom;
         this.p1=new Point(x,y);
         this.cote=cote;
